# Projet_Examen/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portail Scolaire API",
        "description": "Documentation de l'API du portail scolaire",
        "contact": {
            "email": "diabydaj@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/eleves": {
            "get": {
                "tags": [
                    "Élèves"
                ],
                "summary": "Liste des élèves",
                "operationId": "3f8e24fb18108f0deb15dfa293eaa7a6",
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Élèves"
                ],
                "summary": "Créer un nouvel élève",
                "operationId": "de63c08dd7e1c2074b962128f9e58127",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "adresse",
                                    "telephone",
                                    "email",
                                    "date_naissance",
                                    "lieu_naissance",
                                    "sexe",
                                    "classe_id",
                                    "parent_nom",
                                    "parent_prenom",
                                    "parent_adresse",
                                    "parent_telephone",
                                    "parent_email"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Élève créé"
                    }
                }
            }
        },
        "/api/eleves/{id}": {
            "get": {
                "tags": [
                    "Élèves"
                ],
                "summary": "Afficher un élève spécifique",
                "operationId": "312c6407a0cc3cdf847ead762db40a7f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "404": {
                        "description": "Élève non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Élèves"
                ],
                "summary": "Mettre à jour un élève",
                "operationId": "f603e8d396b6d13b654859868702fdc8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "adresse",
                                    "telephone",
                                    "email",
                                    "date_naissance",
                                    "lieu_naissance",
                                    "sexe",
                                    "classe_id"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Élève mis à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Élèves"
                ],
                "summary": "Supprimer un élève",
                "operationId": "370f2b1523fb59334c25ce91085a705b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Élève supprimé"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}